doc and unnecessary cast

+++ core/src/main/java/dk/alexandra/fresco/framework/network/socket/Connector.java
@@ -137,7 +137,7 @@ public class Connector implements NetworkConnector {
           } catch (ConnectException e) {
             // A ConnectionException is expected if the opposing side is not listening for our
             // connection attempt yet. We ignore this and try again.
-            Thread.sleep(Math.min((long)  1 << attempts++, 5_000));
+            Thread.sleep(Math.min(1 << attempts++, 5_000));
             // This should probably not busy-wait for each party
           }
         }
